+++ src/tika_client/__init__.py
@@ -1,3 +1,6 @@
 from tika_client.client import TikaClient
+from tika_client.data_models import DublinCoreKey
+from tika_client.data_models import TikaKey
+from tika_client.data_models import XmpKey
 
-__all__ = ["TikaClient"]
+__all__ = ["TikaClient", "TikaKey", "XmpKey", "DublinCoreKey"]