+++ tests/test_resource_metadata.py
@@ -9,7 +9,6 @@ from pytest_httpx import HTTPXMock
 from tests.conftest import SAMPLE_DIR
 from tests.conftest import TIKA_URL
 from tika_client.client import TikaClient
-from tika_client.data_models import ParsedDocument
 
 
 class TestMetadataResource:
@@ -17,7 +16,6 @@ class TestMetadataResource:
         test_file = SAMPLE_DIR / "sample.docx"
         resp = tika_client.metadata.from_file(test_file, magic.from_file(str(test_file), mime=True))
 
-        assert isinstance(resp, ParsedDocument)
         assert resp.type == "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
         assert resp.created is None
 
@@ -25,7 +23,6 @@ class TestMetadataResource:
         test_file = SAMPLE_DIR / "sample.docx"
         resp = tika_client.metadata.from_file(test_file)
 
-        assert isinstance(resp, ParsedDocument)
         assert resp.type == "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
         assert resp.created is None
 
@@ -33,7 +30,6 @@ class TestMetadataResource:
         test_file = SAMPLE_DIR / "microsoft-sample.docx"
         resp = tika_client.metadata.from_file(test_file, magic.from_file(str(test_file), mime=True))
 
-        assert isinstance(resp, ParsedDocument)
         assert resp.type == "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
         assert resp.created == datetime(year=2023, month=5, day=17, hour=16, minute=41, tzinfo=timezone.utc)
         assert resp.modified == datetime(year=2023, month=5, day=17, hour=16, minute=44, tzinfo=timezone.utc)
@@ -42,7 +38,6 @@ class TestMetadataResource:
         test_file = SAMPLE_DIR / "sample.odt"
         resp = tika_client.metadata.from_file(test_file, magic.from_file(str(test_file), mime=True))
 
-        assert isinstance(resp, ParsedDocument)
         assert resp.type == "application/vnd.oasis.opendocument.text"
         assert resp.data["generator"] == "LibreOfficeDev/6.0.5.2$Linux_X86_64 LibreOffice_project/"
         assert resp.created is None
@@ -51,7 +46,6 @@ class TestMetadataResource:
         test_file = SAMPLE_DIR / "sample.doc"
         resp = tika_client.metadata.from_file(test_file, magic.from_file(str(test_file), mime=True))
 
-        assert isinstance(resp, ParsedDocument)
         assert resp.type == "application/msword"
         assert resp.language == "en"