+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Handling of filenames in the `Content-Disposition` header with non-ASCII characters
 
+### Changed
+
+- All endpoints now return a `TikaResponse`, which will have many of the common keys parsed into Python
+  native data types where possible, based on the list [from the Tika wiki](https://cwiki.apache.org/confluence/display/TIKA/Metadata+Overview).
+  If a key is not in the response, the value will be `None`
+
 ## [0.1.1] - 2023-06-18
 
 ### Fixed